+++ tests/capping/test_BaseCappingTransformer.py
@@ -433,7 +433,7 @@ class GenericCappingTransformTests(GenericTransformTests):
             return
 
         transformer.fit(df)
-        transformer = _handle_from_json(transformer, from_json)
+        transformer = _handle_from_json(transformer, from_json=from_json)
 
         df_transformed = transformer.transform(df)
 
@@ -482,7 +482,7 @@ class GenericCappingTransformTests(GenericTransformTests):
             return
 
         transformer.fit(df)
-        transformer = _handle_from_json(transformer, from_json)
+        transformer = _handle_from_json(transformer, from_json=from_json)
 
         learnt_values = getattr(transformer, fit_value)
 
@@ -524,7 +524,7 @@ class GenericCappingTransformTests(GenericTransformTests):
             return
 
         transformer.fit(df)
-        transformer = _handle_from_json(transformer, from_json)
+        transformer = _handle_from_json(transformer, from_json=from_json)
 
         # convert column to non-numeric
         df = nw.from_native(df)
@@ -591,7 +591,7 @@ class GenericCappingTransformTests(GenericTransformTests):
             return
 
         transformer.fit(df)
-        transformer = _handle_from_json(transformer, from_json)
+        transformer = _handle_from_json(transformer, from_json=from_json)
         transformer.quantile_capping_values = {}
 
         with pytest.raises(
@@ -621,7 +621,7 @@ class GenericCappingTransformTests(GenericTransformTests):
             return
 
         transformer.fit(df)
-        transformer = _handle_from_json(transformer, from_json)
+        transformer = _handle_from_json(transformer, from_json=from_json)
         transformer.capping_values = {}
 
         with pytest.raises(
@@ -681,7 +681,7 @@ class GenericCappingTransformTests(GenericTransformTests):
         transformer.quantile_capping_values = {"a": [1, 4]}
         transformer._replacement_values = {}
 
-        transformer = _handle_from_json(transformer, from_json)
+        transformer = _handle_from_json(transformer, from_json=from_json)
 
         with pytest.raises(
             ValueError,
@@ -711,7 +711,7 @@ class GenericCappingTransformTests(GenericTransformTests):
             return
 
         transformer.fit(df)
-        transformer = _handle_from_json(transformer, from_json)
+        transformer = _handle_from_json(transformer, from_json=from_json)
 
         transformer2 = uninitialized_transformers[self.transformer_name](**args)
 
@@ -719,7 +719,7 @@ class GenericCappingTransformTests(GenericTransformTests):
             return
 
         transformer2.fit(df)
-        transformer2 = _handle_from_json(transformer2, from_json)
+        transformer2 = _handle_from_json(transformer2, from_json=from_json)
 
         transformer2.transform(df)
 
@@ -764,7 +764,7 @@ class GenericCappingTransformTests(GenericTransformTests):
             return
 
         transformer.fit(df)
-        transformer = _handle_from_json(transformer, from_json)
+        transformer = _handle_from_json(transformer, from_json=from_json)
 
         df_transformed = transformer.transform(df)