+++ tests/dates/test_DateTimeInfoExtractor.py
@@ -293,7 +293,7 @@ class TestTransform(
             include=["timeofmonth", "timeofyear", "dayofweek", "timeofday"],
         )
 
-        transformer = _handle_from_json(transformer, from_json)
+        transformer = _handle_from_json(transformer, from_json=from_json)
 
         transformed = transformer.transform(df.to_native())
 
@@ -546,7 +546,7 @@ class TestTransform(
             include=["timeofmonth"],
         )
 
-        transformer = _handle_from_json(transformer, from_json)
+        transformer = _handle_from_json(transformer, from_json=from_json)
 
         transformed = transformer.transform(df.to_native())
 
@@ -684,7 +684,7 @@ class TestTransform(
             },
         )
 
-        transformer = _handle_from_json(transformer, from_json)
+        transformer = _handle_from_json(transformer, from_json=from_json)
 
         transformed = transformer.transform(df)
         expected = nw.from_native(df).clone()