+++ tests/mapping/test_MappingTransformer.py
@@ -88,7 +88,7 @@ class TestTransform(BaseMappingTransformerTransformTests, ReturnNativeTests):
 
         x = MappingTransformer(mappings=mapping, return_dtypes=return_dtypes)
 
-        x = _handle_from_json(x, from_json)
+        x = _handle_from_json(x, from_json=from_json)
 
         df_transformed = x.transform(df)
 
@@ -122,7 +122,7 @@ class TestTransform(BaseMappingTransformerTransformTests, ReturnNativeTests):
 
         x = MappingTransformer(mappings=mapping, return_dtypes=return_dtypes)
 
-        x = _handle_from_json(x, from_json)
+        x = _handle_from_json(x, from_json=from_json)
 
         df_transformed = x.transform(df)
 
@@ -175,7 +175,7 @@ class TestTransform(BaseMappingTransformerTransformTests, ReturnNativeTests):
         df = d.create_df_1(library=library)
         x = MappingTransformer(mappings=mapping, return_dtypes=return_dtypes)
 
-        x = _handle_from_json(x, from_json)
+        x = _handle_from_json(x, from_json=from_json)
 
         df = x.transform(df)
 
@@ -202,7 +202,7 @@ class TestTransform(BaseMappingTransformerTransformTests, ReturnNativeTests):
 
         x = MappingTransformer(mappings=mapping, return_dtypes=return_dtypes)
 
-        x = _handle_from_json(x, from_json)
+        x = _handle_from_json(x, from_json=from_json)
 
         df = x.transform(df)
 
@@ -235,7 +235,7 @@ class TestTransform(BaseMappingTransformerTransformTests, ReturnNativeTests):
             verbose=True,
         )
 
-        x = _handle_from_json(x, from_json)
+        x = _handle_from_json(x, from_json=from_json)
 
         with pytest.warns(
             UserWarning,
@@ -268,7 +268,7 @@ class TestTransform(BaseMappingTransformerTransformTests, ReturnNativeTests):
             verbose=True,
         )
 
-        x = _handle_from_json(x, from_json)
+        x = _handle_from_json(x, from_json=from_json)
 
         with pytest.warns(
             UserWarning,
@@ -371,7 +371,7 @@ class TestTransform(BaseMappingTransformerTransformTests, ReturnNativeTests):
 
         transformer = MappingTransformer(mappings=mapping, return_dtypes=return_dtypes)
 
-        transformer = _handle_from_json(transformer, from_json)
+        transformer = _handle_from_json(transformer, from_json=from_json)
 
         df_transformed = transformer.transform(df)
 
@@ -393,7 +393,7 @@ class TestTransform(BaseMappingTransformerTransformTests, ReturnNativeTests):
             verbose=True,
         )
 
-        transformer = _handle_from_json(transformer, from_json)
+        transformer = _handle_from_json(transformer, from_json=from_json)
 
         with pytest.warns(
             UserWarning,
@@ -419,7 +419,7 @@ class TestTransform(BaseMappingTransformerTransformTests, ReturnNativeTests):
 
         transformer = MappingTransformer(mappings=mapping, return_dtypes=return_dtypes)
 
-        transformer = _handle_from_json(transformer, from_json)
+        transformer = _handle_from_json(transformer, from_json=from_json)
 
         # Guidance from pytest to use followiong syntax rather than pytest.warns(None)
         try: