+++ tubular/base.py
@@ -159,6 +159,7 @@ class BaseTransformer(BaseEstimator, TransformerMixin):
             NonEmptyListOfStrs,
             str,
         ],
+        *,
         copy: bool = False,
         verbose: bool = False,
         return_native: bool = True,
@@ -425,7 +426,7 @@ class BaseTransformer(BaseEstimator, TransformerMixin):
         return X.with_columns(_temporary_response=y)
 
     @beartype
-    def _process_return_native(self, return_native_override: Optional[bool]) -> bool:
+    def _process_return_native(self, *, return_native_override: Optional[bool]) -> bool:
         """Determine whether to override return_native attr.
 
         Parameters
@@ -459,6 +460,7 @@ class BaseTransformer(BaseEstimator, TransformerMixin):
     def transform(
         self,
         X: DataFrame,
+        *,
         return_native_override: Optional[bool] = None,
     ) -> DataFrame:
         """Check data before child transform.
@@ -504,7 +506,9 @@ class BaseTransformer(BaseEstimator, TransformerMixin):
         ```
 
         """
-        return_native = self._process_return_native(return_native_override)
+        return_native = self._process_return_native(
+            return_native_override=return_native_override
+        )
 
         X = _convert_dataframe_to_narwhals(X)
 
@@ -517,7 +521,7 @@ class BaseTransformer(BaseEstimator, TransformerMixin):
         if self.verbose:
             print("BaseTransformer.transform() called")
 
-        return _return_narwhals_or_native_dataframe(X, return_native)
+        return _return_narwhals_or_native_dataframe(X, return_native=return_native)
 
     def check_is_fitted(self, attribute: str) -> None:
         """Check if particular attributes are on the object.
@@ -654,6 +658,7 @@ class DataFrameMethodTransformer(DropOriginalMixin, BaseTransformer):
         pd_method_name: str,
         columns: Optional[Union[list[str], str]],
         pd_method_kwargs: Optional[GenericKwargs] = None,
+        *,
         drop_original: bool = False,
         **kwargs: Optional[bool],
     ) -> None:
@@ -758,6 +763,6 @@ class DataFrameMethodTransformer(DropOriginalMixin, BaseTransformer):
         # Drop original columns if self.drop_original is True
         return DropOriginalMixin.drop_original_column(
             X,
-            self.drop_original,
             self.columns,
+            drop_original=self.drop_original,
         )