+++ tubular/dates.py
@@ -90,6 +90,7 @@ class BaseGenericDateTransformer(
         self,
         columns: Union[list[str], str],
         new_column_name: str,
+        *,
         drop_original: bool = False,
         **kwargs: Optional[bool],
     ) -> None:
@@ -193,6 +194,7 @@ class BaseGenericDateTransformer(
     def check_columns_are_date_or_datetime(
         self,
         X: DataFrame,
+        *,
         datetime_only: bool,
     ) -> None:
         """Check types of provided columns.
@@ -286,6 +288,7 @@ class BaseGenericDateTransformer(
     def transform(
         self,
         X: DataFrame,
+        *,
         datetime_only: bool = False,
         return_native_override: Optional[bool] = None,
     ) -> DataFrame:
@@ -341,7 +344,9 @@ class BaseGenericDateTransformer(
         ```
 
         """
-        return_native = self._process_return_native(return_native_override)
+        return_native = self._process_return_native(
+            return_native_override=return_native_override
+        )
 
         X = super().transform(X, return_native_override=False)
 
@@ -349,7 +354,7 @@ class BaseGenericDateTransformer(
 
         self.check_columns_are_date_or_datetime(X, datetime_only=datetime_only)
 
-        return _return_narwhals_or_native_dataframe(X, return_native)
+        return _return_narwhals_or_native_dataframe(X, return_native=return_native)
 
 
 @register
@@ -400,6 +405,7 @@ class BaseDatetimeTransformer(BaseGenericDateTransformer):
         self,
         columns: Union[list[str], str],
         new_column_name: str,
+        *,
         drop_original: bool = False,
         **kwargs: Optional[bool],
     ) -> None:
@@ -431,6 +437,7 @@ class BaseDatetimeTransformer(BaseGenericDateTransformer):
     def transform(
         self,
         X: DataFrame,
+        *,
         return_native_override: Optional[bool] = None,
     ) -> DataFrame:
         """Check types of selected columns in provided data.
@@ -482,13 +489,15 @@ class BaseDatetimeTransformer(BaseGenericDateTransformer):
         ```
 
         """
-        return_native = self._process_return_native(return_native_override)
+        return_native = self._process_return_native(
+            return_native_override=return_native_override
+        )
 
         X = _convert_dataframe_to_narwhals(X)
 
         X = super().transform(X, datetime_only=True, return_native_override=False)
 
-        return _return_narwhals_or_native_dataframe(X, return_native)
+        return _return_narwhals_or_native_dataframe(X, return_native=return_native)
 
 
 class DateDifferenceUnitsOptions(str, Enum):
@@ -575,8 +584,9 @@ class DateDifferenceTransformer(BaseGenericDateTransformer):
         columns: ListOfTwoStrs,
         new_column_name: str,
         units: DateDifferenceUnitsOptionsStr = "D",
-        drop_original: bool = False,
         custom_days_divider: Optional[int] = None,
+        *,
+        drop_original: bool = False,
         **kwargs: bool,
     ) -> None:
         """Initialise class instance.
@@ -750,12 +760,12 @@ class DateDifferenceTransformer(BaseGenericDateTransformer):
         # Drop original columns if self.drop_original is True
         X = DropOriginalMixin.drop_original_column(
             X,
-            self.drop_original,
             self.columns,
+            drop_original=self.drop_original,
             return_native=False,
         )
 
-        return _return_narwhals_or_native_dataframe(X, self.return_native)
+        return _return_narwhals_or_native_dataframe(X, return_native=self.return_native)
 
 
 @register
@@ -966,6 +976,7 @@ class BetweenDatesTransformer(BaseGenericDateTransformer):
         self,
         columns: ListOfThreeStrs,
         new_column_name: str,
+        *,
         drop_original: bool = False,
         lower_inclusive: bool = True,
         upper_inclusive: bool = True,
@@ -1130,8 +1141,8 @@ class BetweenDatesTransformer(BaseGenericDateTransformer):
         # Drop original columns if self.drop_original is True
         return DropOriginalMixin.drop_original_column(
             X,
-            self.drop_original,
             self.columns,
+            drop_original=self.drop_original,
         )
 
 
@@ -1271,6 +1282,7 @@ class DatetimeInfoExtractor(BaseDatetimeTransformer):
         columns: Union[str, list[str]],
         include: Optional[Union[DatetimeInfoOptionList, DatetimeInfoOptionStr]] = None,
         datetime_mappings: Optional[dict[DatetimeInfoOptionStr, dict[int, str]]] = None,
+        *,
         drop_original: Optional[bool] = False,
         **kwargs: Union[str, bool],
     ) -> None:
@@ -1545,12 +1557,12 @@ class DatetimeInfoExtractor(BaseDatetimeTransformer):
         # Drop original columns if self.drop_original is True
         X = DropOriginalMixin.drop_original_column(
             X,
-            self.drop_original,
             self.columns,
             return_native=False,
+            drop_original=self.drop_original,
         )
 
-        return _return_narwhals_or_native_dataframe(X, self.return_native)
+        return _return_narwhals_or_native_dataframe(X, return_native=self.return_native)
 
 
 class DatetimeComponentOptions(str, Enum):
@@ -1793,7 +1805,7 @@ class DatetimeComponentExtractor(BaseDatetimeTransformer):
             **transform_dict,
         )
 
-        return _return_narwhals_or_native_dataframe(X, self.return_native)
+        return _return_narwhals_or_native_dataframe(X, return_native=self.return_native)
 
 
 class DatetimeSinusoidUnitsOptions(str, Enum):
@@ -1920,6 +1932,7 @@ class DatetimeSinusoidCalculator(BaseDatetimeTransformer):
             dict[str, DatetimeSinusoidUnitsOptionStr],
         ],
         period: Union[NumberNotBool, dict[str, NumberNotBool]] = 2 * np.pi,
+        *,
         drop_original: bool = False,
         **kwargs: Union[bool, str],
     ) -> None:
@@ -2050,6 +2063,7 @@ class DatetimeSinusoidCalculator(BaseDatetimeTransformer):
     def transform(
         self,
         X: DataFrame,
+        *,
         return_native_override: Optional[bool] = None,
     ) -> DataFrame:
         """Transform - creates column containing sine or cosine of another datetime column.
@@ -2104,7 +2118,9 @@ class DatetimeSinusoidCalculator(BaseDatetimeTransformer):
 
         """
         X = _convert_dataframe_to_narwhals(X)
-        return_native = self._process_return_native(return_native_override)
+        return_native = self._process_return_native(
+            return_native_override=return_native_override
+        )
 
         X = super().transform(X, return_native_override=False)
 
@@ -2150,11 +2166,11 @@ class DatetimeSinusoidCalculator(BaseDatetimeTransformer):
         # Drop original columns if self.drop_original is True
         X = DropOriginalMixin.drop_original_column(
             X,
-            self.drop_original,
             self.columns,
+            drop_original=self.drop_original,
             return_native=False,
         )
-        return _return_narwhals_or_native_dataframe(X, return_native)
+        return _return_narwhals_or_native_dataframe(X, return_native=return_native)
 
 
 # DEPRECATED TRANSFORMERS
@@ -2220,6 +2236,7 @@ class DateDiffLeapYearTransformer(BaseGenericDateTransformer):
         columns: ListOfTwoStrs,
         new_column_name: str,
         missing_replacement: Optional[Union[float, int, str]] = None,
+        *,
         drop_original: bool = False,
         **kwargs: bool,
     ) -> None:
@@ -2325,8 +2342,8 @@ class DateDiffLeapYearTransformer(BaseGenericDateTransformer):
         # Drop original columns if self.drop_original is True
         return DropOriginalMixin.drop_original_column(
             X,
-            self.drop_original,
             self.columns,
+            drop_original=self.drop_original,
         )
 
 
@@ -2411,6 +2428,7 @@ class SeriesDtMethodTransformer(BaseDatetimeTransformer):
             str,
         ],
         pd_method_kwargs: Optional[GenericKwargs] = None,
+        *,
         drop_original: bool = False,
         **kwargs: Optional[bool],
     ) -> None:
@@ -2513,6 +2531,6 @@ class SeriesDtMethodTransformer(BaseDatetimeTransformer):
         # Drop original columns if self.drop_original is True
         return DropOriginalMixin.drop_original_column(
             X,
-            self.drop_original,
             self.columns,
+            drop_original=self.drop_original,
         )